ref: mcp_server code modularization
