+++ modules/cli/cmd/mcp_server.go
@@ -18,23 +18,16 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"net/http"
-	"net/http/httptest"
-	"strconv"
 	"strings"
 	"time"
 
 	"github.com/mark3labs/mcp-go/mcp"
 	"github.com/mark3labs/mcp-go/server"
 	"github.com/rs/zerolog/log"
-	"github.com/sosodev/duration"
 	"github.com/spf13/cobra"
 
 	"github.com/kubetail-org/kubetail/modules/shared/k8shelpers"
 	"github.com/kubetail-org/kubetail/modules/shared/logs"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes"
-	"k8s.io/kubectl/pkg/proxy"
 )
 
 const mcpServerHelp = `
@@ -66,356 +59,6 @@ Add this to your claude_desktop_config.json:
   }
 }
 `
-
-// KubernetesLogsSearchArgs represents the arguments for the kubernetes_logs_search MCP tool
-type KubernetesLogsSearchArgs struct {
-	// Natural language query (required)
-	Query string `json:"query" jsonschema:"required,description=Natural language query for logs (e.g. 'nginx errors', 'API timeouts in production')"`
-
-	// Basic Kubernetes filters
-	Namespace string `json:"namespace,omitempty" jsonschema:"description=Kubernetes namespace to search (optional, will search default if not specified)"`
-	Pod       string `json:"pod,omitempty" jsonschema:"description=Specific pod name or pattern to search (optional, supports wildcards like 'nginx-*')"`
-	Container string `json:"container,omitempty" jsonschema:"description=Specific container name to search (optional)"`
-
-	// Advanced infrastructure filters
-	Region string `json:"region,omitempty" jsonschema:"description=Filter by cloud region (e.g. 'us-east-1,us-west-2')"`
-	Zone   string `json:"zone,omitempty" jsonschema:"description=Filter by availability zone (e.g. 'us-east-1a,us-east-1b')"`
-	Arch   string `json:"arch,omitempty" jsonschema:"description=Filter by CPU architecture (e.g. 'amd64,arm64')"`
-	OS     string `json:"os,omitempty" jsonschema:"description=Filter by operating system (e.g. 'linux,windows')"`
-	Node   string `json:"node,omitempty" jsonschema:"description=Filter by specific node names"`
-
-	// Content filtering
-	Grep string `json:"grep,omitempty" jsonschema:"description=Text pattern to grep for in log lines (regex supported, processed server-side)"`
-
-	// Time-based filtering
-	Since string `json:"since,omitempty" jsonschema:"description=Show logs since this time (ISO timestamp or duration like 'PT30M')"`
-	Until string `json:"until,omitempty" jsonschema:"description=Show logs until this time (ISO timestamp or duration like 'PT10M')"`
-
-	// Output control
-	Tail   int64 `json:"tail,omitempty" jsonschema:"description=Show last N log entries (default: 100)"`
-	Follow bool  `json:"follow,omitempty" jsonschema:"description=Stream new log entries in real-time"`
-}
-
-// Response structures for AI tool consumption
-type LogEntryResponse struct {
-	Timestamp string            `json:"timestamp"`
-	Message   string            `json:"message"`
-	Source    logs.LogSource    `json:"source"`
-	Level     string            `json:"level,omitempty"`    // Detected log level
-	Tags      []string          `json:"tags,omitempty"`     // Detected tags/labels
-	Metadata  map[string]string `json:"metadata,omitempty"` // Additional context
-}
-
-type LogSearchResponse struct {
-	Query         string                 `json:"query"`
-	Summary       LogSummary             `json:"summary"`
-	Entries       []LogEntryResponse     `json:"entries"`
-	Filters       AppliedFilters         `json:"filters"`
-	Processing    ProcessingInfo         `json:"processing"`
-	InfraInsights InfrastructureInsights `json:"infrastructure_insights"`
-}
-
-type LogSummary struct {
-	TotalEntries  int            `json:"total_entries"`
-	TimeRange     TimeRange      `json:"time_range"`
-	SourceSummary SourceSummary  `json:"sources"`
-	LevelCounts   map[string]int `json:"level_counts,omitempty"`
-	TopKeywords   []string       `json:"top_keywords,omitempty"`
-	ErrorCount    int            `json:"error_count,omitempty"`
-	WarningCount  int            `json:"warning_count,omitempty"`
-}
-
-type TimeRange struct {
-	Start string `json:"start"`
-	End   string `json:"end"`
-}
-
-type SourceSummary struct {
-	Namespaces []string `json:"namespaces"`
-	Pods       []string `json:"pods"`
-	Containers []string `json:"containers"`
-}
-
-type AppliedFilters struct {
-	Namespace string   `json:"namespace,omitempty"`
-	Pod       string   `json:"pod,omitempty"`
-	Container string   `json:"container,omitempty"`
-	Region    []string `json:"region,omitempty"`
-	Zone      []string `json:"zone,omitempty"`
-	Arch      []string `json:"arch,omitempty"`
-	OS        []string `json:"os,omitempty"`
-	Node      []string `json:"node,omitempty"`
-	Grep      string   `json:"grep,omitempty"`
-	Since     string   `json:"since,omitempty"`
-	Until     string   `json:"until,omitempty"`
-	Tail      int64    `json:"tail,omitempty"`
-}
-
-type ProcessingInfo struct {
-	LogFetcher       string `json:"log_fetcher"`        // "AgentLogFetcher" or "KubeLogFetcher"
-	ServerSideGrep   bool   `json:"server_side_grep"`   // Whether grep was processed server-side
-	ProcessingTimeMs int64  `json:"processing_time_ms"` // How long the query took
-	SourcesScanned   int    `json:"sources_scanned"`    // Number of log sources checked
-}
-
-// InfrastructureInsights captures Kubetail's infrastructure-aware capabilities
-type InfrastructureInsights struct {
-	AdvancedFiltering   bool     `json:"advanced_filtering"`    // Whether infrastructure-based filters were used (cloud region/os/ cpu architecture)
-	ServerSideFiltering bool     `json:"server_side_filtering"` // Whether filtering happened on agents vs client
-	ProcessingMethod    string   `json:"processing_method"`     // Description of log processing approach
-	AppliedFilters      []string `json:"applied_filters"`       // List of applied filters
-	InfrastructureSpan  string   `json:"infrastructure_span"`   // Description of infrastructure coverage
-}
-
-// ClusterAPIProxyLogFetcher implements LogFetcher using Kubernetes service proxy to reach Cluster API
-// This provides automatic service discovery without requiring manual URL configuration
-type ClusterAPIProxyLogFetcher struct {
-	cm          k8shelpers.ConnectionManager
-	kubeContext string
-	namespace   string
-	serviceName string
-	client      *http.Client
-}
-
-// NewClusterAPIProxyLogFetcher initializes a ClusterAPIProxyLogFetcher for accessing the cluster API service via Kubernetes proxy.
-func NewClusterAPIProxyLogFetcher(cm k8shelpers.ConnectionManager, kubeContext string) (*ClusterAPIProxyLogFetcher, error) {
-	// Try to find kubetail-cluster-api service in kubetail-system namespace first
-	namespace := "kubetail-system"
-	serviceName := "kubetail-cluster-api"
-
-	if clientset, err := cm.GetOrCreateClientset(kubeContext); err == nil {
-		// Check if the cluster API service exists in kubetail-system
-		if _, err := clientset.CoreV1().Services(namespace).Get(context.TODO(), serviceName, metav1.GetOptions{}); err != nil {
-			log.Debug().Str("namespace", namespace).Str("service", serviceName).Msg("kubetail-cluster-api service not found in kubetail-system, trying default namespace")
-			namespace = "default"
-			// Check if it exists in default namespace
-			if _, err := clientset.CoreV1().Services(namespace).Get(context.TODO(), serviceName, metav1.GetOptions{}); err != nil {
-				return nil, fmt.Errorf("kubetail-cluster-api service not found in %s or %s namespaces", "kubetail-system", "default")
-			}
-		}
-	}
-
-	log.Debug().Str("namespace", namespace).Str("service", serviceName).Msg("Found kubetail-cluster-api service")
-
-	return &ClusterAPIProxyLogFetcher{
-		cm:          cm,
-		kubeContext: kubeContext,
-		namespace:   namespace,
-		serviceName: serviceName,
-		client:      &http.Client{Timeout: 30 * time.Second},
-	}, nil
-}
-
-// StreamForward returns a channel of LogRecords in chronological order
-func (f *ClusterAPIProxyLogFetcher) StreamForward(ctx context.Context, source logs.LogSource, opts logs.FetcherOptions) (<-chan logs.LogRecord, error) {
-	return f.streamLogs(ctx, source, opts, false) // false = forward direction
-}
-
-// StreamBackward returns a channel of LogRecords in reverse chronological order
-func (f *ClusterAPIProxyLogFetcher) StreamBackward(ctx context.Context, source logs.LogSource, opts logs.FetcherOptions) (<-chan logs.LogRecord, error) {
-	return f.streamLogs(ctx, source, opts, true) // true = backward direction
-}
-
-// streamLogs fetches log records from the cluster API proxy using a GraphQL query and streams them to a channel.
-func (f *ClusterAPIProxyLogFetcher) streamLogs(ctx context.Context, source logs.LogSource, opts logs.FetcherOptions, backward bool) (<-chan logs.LogRecord, error) {
-
-	outCh := make(chan logs.LogRecord)
-
-	go func() {
-		defer close(outCh)
-
-		handler, token, err := f.getKubernetesServiceProxyHandler(ctx)
-		if err != nil {
-			log.Debug().Err(err).Msg("Could not create Kubernetes service proxy handler")
-			return
-		}
-
-		query := f.buildLogRecordsQuery(source, opts, backward, 100)
-
-		reqBody := fmt.Sprintf(`{"query": %s}`, strconv.Quote(query))
-
-		proxyPath := fmt.Sprintf("/api/v1/namespaces/%s/services/%s:http/proxy/graphql", f.namespace, f.serviceName)
-		req, err := http.NewRequestWithContext(ctx, "POST", proxyPath, strings.NewReader(reqBody))
-		if err != nil {
-			log.Debug().Err(err).Msg("Could not create proxy request for log records")
-			return
-		}
-
-		req.Header.Set("Content-Type", "application/json")
-		req.Header.Set("X-Forwarded-Authorization", fmt.Sprintf("Bearer %s", token))
-
-		// Execute request through proxy
-		rec := httptest.NewRecorder()
-		handler.ServeHTTP(rec, req)
-
-		if rec.Code != 200 {
-			log.Debug().Int("status", rec.Code).Msg("Cluster API proxy request returned error status")
-			return
-		}
-
-		// Parse the GraphQL response
-		var response struct {
-			Data struct {
-				LogRecordsFetch struct {
-					Records []struct {
-						Timestamp string `json:"timestamp"`
-						Message   string `json:"message"`
-					} `json:"records"`
-				} `json:"logRecordsFetch"`
-			} `json:"data"`
-			Errors []struct {
-				Message string `json:"message"`
-			} `json:"errors"`
-		}
-
-		if err := json.Unmarshal(rec.Body.Bytes(), &response); err != nil {
-			log.Debug().Err(err).Msg("Could not parse GraphQL response from cluster API")
-			return
-		}
-
-		// Check for GraphQL errors
-		if len(response.Errors) > 0 {
-			for _, gqlErr := range response.Errors {
-				log.Debug().Str("error", gqlErr.Message).Msg("Cluster API GraphQL query error")
-			}
-			return
-		}
-
-		// Process log records
-		for _, record := range response.Data.LogRecordsFetch.Records {
-			// Parse timestamp
-			timestamp, err := time.Parse(time.RFC3339Nano, record.Timestamp)
-			if err != nil {
-				// Fallback to RFC3339 if nano parsing fails
-				timestamp, err = time.Parse(time.RFC3339, record.Timestamp)
-				if err != nil {
-					log.Debug().Err(err).Str("timestamp", record.Timestamp).Msg("Could not parse timestamp")
-					timestamp = time.Now() // Use current time as fallback
-				}
-			}
-
-			logRecord := logs.LogRecord{
-				Message:   record.Message,
-				Timestamp: timestamp,
-				Source:    source,
-			}
-
-			select {
-			case <-ctx.Done():
-				return
-			case outCh <- logRecord:
-				// Successfully sent record
-			}
-		}
-	}()
-
-	return outCh, nil
-}
-
-// getKubernetesServiceProxyHandler returns an HTTP handler and token for proxying requests to the cluster API service.
-func (f *ClusterAPIProxyLogFetcher) getKubernetesServiceProxyHandler(ctx context.Context) (http.Handler, string, error) {
-
-	restConfig, err := f.cm.GetOrCreateRestConfig(f.kubeContext)
-	if err != nil {
-		return nil, "", err
-	}
-
-	// Create proxy handler using kubectl proxy functionality
-	handler, err := proxy.NewProxyHandler("/", nil, restConfig, 0, false)
-	if err != nil {
-		return nil, "", err
-	}
-
-	// Get service account token for authentication
-	clientset, err := f.cm.GetOrCreateClientset(f.kubeContext)
-	if err != nil {
-		return nil, "", err
-	}
-
-	// Try to create service account token for cluster API access
-	// First try kubetail-mcp service account, fallback to default
-	var token string
-	var tokenErr error
-
-	serviceAccounts := []string{"kubetail-cli", "kubetail-mcp", "default"}
-
-	for _, sa := range serviceAccounts {
-		sat, saErr := k8shelpers.NewServiceAccountToken(ctx, clientset, f.namespace, sa, nil)
-		if saErr != nil {
-			log.Debug().Err(saErr).Str("service_account", sa).Msg("Could not create service account token helper")
-			continue
-		}
-
-		token, tokenErr = sat.Token(ctx)
-		if tokenErr != nil {
-			log.Debug().Err(tokenErr).Str("service_account", sa).Msg("Could not get token from service account")
-			continue
-		}
-
-		log.Debug().Str("service_account", sa).Msg("Successfully obtained service account token")
-		break
-	}
-
-	if token == "" || tokenErr != nil {
-		return nil, "", fmt.Errorf("could not obtain service account token from any of %v: %w", serviceAccounts, tokenErr)
-	}
-
-	return handler, token, nil
-}
-
-// buildLogRecordsQuery constructs a GraphQL query string for fetching log records from the cluster API.
-func (f *ClusterAPIProxyLogFetcher) buildLogRecordsQuery(source logs.LogSource, opts logs.FetcherOptions, backward bool, limit int) string {
-	var queryParts []string
-
-	// Build sources array as strings (like "namespace:pod/container")
-	sourceStr := source.Namespace + ":"
-	if source.PodName != "" {
-		sourceStr += source.PodName
-		if source.ContainerName != "" {
-			sourceStr += "/" + source.ContainerName
-		}
-	} else {
-		sourceStr += "*" // All pods if no specific pod
-	}
-
-	queryParts = append(queryParts, fmt.Sprintf(`sources: ["%s"]`, sourceStr))
-
-	if backward {
-		queryParts = append(queryParts, `mode: HEAD`)
-	} else {
-		queryParts = append(queryParts, `mode: TAIL`)
-	}
-
-	if opts.Grep != "" {
-		queryParts = append(queryParts, fmt.Sprintf(`grep: "%s"`, opts.Grep))
-	}
-
-	if !opts.StartTime.IsZero() {
-		queryParts = append(queryParts, fmt.Sprintf(`since: "%s"`, opts.StartTime.Format(time.RFC3339)))
-	}
-
-	if !opts.StopTime.IsZero() {
-		queryParts = append(queryParts, fmt.Sprintf(`until: "%s"`, opts.StopTime.Format(time.RFC3339)))
-	}
-
-	queryParts = append(queryParts, fmt.Sprintf(`limit: %d`, limit))
-
-	// Build the complete query
-	queryArgs := strings.Join(queryParts, ", ")
-
-	query := fmt.Sprintf(`
-		query {
-			logRecordsFetch(%s) {
-				records {
-					timestamp
-					message
-				}
-			}
-		}
-	`, queryArgs)
-
-	return query
-}
 
 // mcpServerCmd represents the mcp-server command
 var mcpServerCmd = &cobra.Command{
@@ -503,501 +146,7 @@ var mcpServerCmd = &cobra.Command{
 			),
 		)
 
-		s.AddTool(tool, func(ctx context.Context, request mcp.CallToolRequest) (*mcp.CallToolResult, error) {
-			log.Info().Str("tool", "kubernetes_logs_search").Msg("Processing log search request")
-
-			args, ok := request.Params.Arguments.(map[string]interface{})
-			if !ok {
-				return mcp.NewToolResultError("invalid arguments format"), nil
-			}
-
-			query, ok := args["query"].(string)
-			if !ok || query == "" {
-				return mcp.NewToolResultError("query parameter is required"), nil
-			}
-
-			// Kubernetes filters
-			namespace, _ := args["namespace"].(string)
-			pod, _ := args["pod"].(string)
-			container, _ := args["container"].(string)
-
-			// Infrastructure filters
-			region, _ := args["region"].(string)
-			zone, _ := args["zone"].(string)
-			arch, _ := args["arch"].(string)
-			osFilter, _ := args["os"].(string)
-			node, _ := args["node"].(string)
-
-			// Content and time filters
-			grep, _ := args["grep"].(string)
-			since, _ := args["since"].(string)
-			until, _ := args["until"].(string)
-
-			// Output control
-			tail, _ := args["tail"].(float64)
-			follow, _ := args["follow"].(bool)
-			format, _ := args["format"].(string)
-
-			// Set defaults
-			if tail == 0 {
-				tail = 100 // default
-			}
-			if format == "" {
-				format = "detailed" // Default to structured JSON for AI consumption
-			}
-
-			sourcePaths, err := buildSourcePaths(namespace, pod, container, cm, kubeContext)
-			if err != nil {
-				return mcp.NewToolResultError(fmt.Sprintf("Failed to build source paths: %v", err)), nil
-			}
-
-			if len(sourcePaths) == 0 {
-				return mcp.NewToolResultText("No matching pods found for the specified criteria"), nil
-			}
-
-			streamOpts := []logs.Option{
-				logs.WithKubeContext(kubeContext),
-				logs.WithTail(int64(tail)),
-				logs.WithFollow(follow),
-			}
-
-			if region != "" {
-				regionList := parseCommaList(region)
-				streamOpts = append(streamOpts, logs.WithRegions(regionList))
-			}
-			if zone != "" {
-				zoneList := parseCommaList(zone)
-				streamOpts = append(streamOpts, logs.WithZones(zoneList))
-			}
-			if arch != "" {
-				archList := parseCommaList(arch)
-				streamOpts = append(streamOpts, logs.WithArches(archList))
-			}
-			if osFilter != "" {
-				osList := parseCommaList(osFilter)
-				streamOpts = append(streamOpts, logs.WithOSes(osList))
-			}
-			if node != "" {
-				nodeList := parseCommaList(node)
-				streamOpts = append(streamOpts, logs.WithNodes(nodeList))
-			}
-
-			if grep != "" {
-				streamOpts = append(streamOpts, logs.WithGrep(grep))
-			}
-
-			if since != "" {
-				sinceTime, err := parseTimeArgMCP(since)
-				if err != nil {
-					return mcp.NewToolResultError(fmt.Sprintf("Invalid since time: %v", err)), nil
-				}
-				streamOpts = append(streamOpts, logs.WithSince(sinceTime))
-			}
-			if until != "" {
-				untilTime, err := parseTimeArgMCP(until)
-				if err != nil {
-					return mcp.NewToolResultError(fmt.Sprintf("Invalid until time: %v", err)), nil
-				}
-				streamOpts = append(streamOpts, logs.WithUntil(untilTime))
-			}
-
-			// Initialize SmartLogFetcher
-			logFetcher, logFetcherType, err := createSmartLogFetcher(cm, kubeContext)
-			if err != nil {
-				return mcp.NewToolResultError(fmt.Sprintf("Failed to create log fetcher: %s", err)), nil
-			}
-			streamOpts = append(streamOpts, logs.WithLogFetcher(logFetcher))
-
-			startTime := time.Now()
-			stream, err := logs.NewStream(ctx, cm, sourcePaths, streamOpts...)
-			if err != nil {
-				return mcp.NewToolResultError(fmt.Sprintf("Failed to create log stream: %v", err)), nil
-			}
-			defer stream.Close()
-
-			if err := stream.Start(ctx); err != nil {
-				return mcp.NewToolResultError(fmt.Sprintf("Failed to start log stream: %v", err)), nil
-			}
-
-			processingTime := time.Since(startTime).Milliseconds()
-
-			appliedFilters := AppliedFilters{
-				Namespace: namespace,
-				Pod:       pod,
-				Container: container,
-				Grep:      grep,
-				Since:     since,
-				Until:     until,
-				Tail:      int64(tail),
-			}
-
-			if region != "" {
-				appliedFilters.Region = parseCommaList(region)
-			}
-			if zone != "" {
-				appliedFilters.Zone = parseCommaList(zone)
-			}
-			if arch != "" {
-				appliedFilters.Arch = parseCommaList(arch)
-			}
-			if osFilter != "" {
-				appliedFilters.OS = parseCommaList(osFilter)
-			}
-			if node != "" {
-				appliedFilters.Node = parseCommaList(node)
-			}
-
-			processingInfo := ProcessingInfo{
-				LogFetcher:       logFetcherType,
-				ServerSideGrep:   grep != "" && logFetcherType == "AgentLogFetcher",
-				ProcessingTimeMs: processingTime,
-				SourcesScanned:   len(sourcePaths),
-			}
-
-			// Handle streaming (follow) requests
-			if follow {
-				var streamingSince time.Time
-
-				// since parameter was provided
-				if since != "" {
-					var err error
-					streamingSince, err = parseTimeArgMCP(since)
-					if err != nil {
-						return mcp.NewToolResultError(fmt.Sprintf("Invalid since time for streaming: %v", err)), nil
-					}
-				} else {
-					// no since parameter, use current time, start collecting from 5 seconds ago
-					streamingSince = time.Now().Add(-5 * time.Second)
-				}
-
-				hasSinceOption := false
-				for i, opt := range streamOpts {
-					// check the option, if it's a WithSince
-					if optStr := fmt.Sprintf("%T", opt); strings.Contains(optStr, "WithSince") {
-						streamOpts[i] = logs.WithSince(streamingSince)
-						hasSinceOption = true
-						break
-					}
-				}
-
-				// add since option if not found
-				if !hasSinceOption {
-					streamOpts = append(streamOpts, logs.WithSince(streamingSince))
-				}
-
-				var initialRecords []logs.LogRecord
-				recordCount := 0
-				recordLimit := 20 // initial batch size
-
-				timeout := time.After(2 * time.Second)
-				collectDone := false
-
-				for !collectDone {
-					select {
-					case record, ok := <-stream.Records():
-						if !ok {
-							collectDone = true
-							break
-						}
-						initialRecords = append(initialRecords, record)
-						recordCount++
-						if recordCount >= recordLimit {
-							collectDone = true
-						}
-					case <-timeout:
-						collectDone = true
-					case <-ctx.Done():
-						return mcp.NewToolResultError("Request cancelled"), nil
-					}
-				}
-
-				// get latest timestamp from logs for next polling request
-				var latestTimestamp time.Time
-				if len(initialRecords) > 0 {
-					for _, record := range initialRecords {
-						if record.Timestamp.After(latestTimestamp) {
-							latestTimestamp = record.Timestamp
-						}
-					}
-				} else {
-					latestTimestamp = time.Now()
-				}
-
-				// 1 millisecond increment to prevent duplicate logs in next poll
-				nextPollTime := latestTimestamp.Add(time.Millisecond)
-
-				response := buildStructuredResponse(query, initialRecords, appliedFilters, processingInfo)
-
-				// meta info
-				if response.InfraInsights.AppliedFilters == nil {
-					response.InfraInsights.AppliedFilters = []string{}
-				}
-				response.InfraInsights.AppliedFilters = append(
-					response.InfraInsights.AppliedFilters,
-					"Real-time log streaming requires repeated calls with the same parameters")
-
-				if len(initialRecords) == 0 {
-					response.Summary.TotalEntries = 0
-					jsonData, err := json.MarshalIndent(response, "", "  ")
-					if err != nil {
-						return mcp.NewToolResultError(fmt.Sprintf("Failed to serialize response: %v", err)), nil
-					}
-
-					// Use current time as the next poll point
-					nextPollStr := time.Now().Format(time.RFC3339)
-					return mcp.NewToolResultText(fmt.Sprintf("No log entries found. For continuous updates, call again with: since=\"%s\"\n\n%s",
-						nextPollStr, string(jsonData))), nil
-				}
-
-				jsonData, err := json.MarshalIndent(response, "", "  ")
-				if err != nil {
-					return mcp.NewToolResultError(fmt.Sprintf("Failed to serialize response: %v", err)), nil
-				}
-
-				nextPollStr := nextPollTime.Format(time.RFC3339)
-
-				return mcp.NewToolResultText(fmt.Sprintf("Found %d log entries. For continuous updates, call again with: since=\"%s\"\n\n%s",
-					len(initialRecords), nextPollStr, string(jsonData))), nil
-			}
-
-			// For non-streaming requests, collect all records
-			var records []logs.LogRecord
-			for record := range stream.Records() {
-				records = append(records, record)
-			}
-
-			if stream.Err() != nil {
-				return mcp.NewToolResultError(fmt.Sprintf("Error while streaming logs: %v", stream.Err())), nil
-			}
-
-			response := buildStructuredResponse(query, records, appliedFilters, processingInfo)
-
-			if len(records) == 0 {
-				return mcp.NewToolResultText("No log entries found matching the specified criteria"), nil
-			}
-
-			switch format {
-			case "raw":
-				var logLines []string
-				for _, record := range records {
-					logLine := fmt.Sprintf("[%s] %s/%s/%s: %s",
-						record.Timestamp.Format("2006-01-02 15:04:05"),
-						record.Source.Namespace,
-						record.Source.PodName,
-						record.Source.ContainerName,
-						record.Message)
-					logLines = append(logLines, logLine)
-				}
-				result := fmt.Sprintf("Found %d log entries:\n\n%s", len(logLines), strings.Join(logLines, "\n"))
-				return mcp.NewToolResultText(result), nil
-
-			case "detailed":
-				// Full structured JSON for advanced AI analysis
-				jsonData, err := json.MarshalIndent(response, "", "  ")
-				if err != nil {
-					return mcp.NewToolResultError(fmt.Sprintf("Failed to serialize structured response: %v", err)), nil
-				}
-				return mcp.NewToolResultText(fmt.Sprintf("Structured log data (JSON):\n\n%s", string(jsonData))), nil
-
-			default: // "summary"
-				// AI-friendly summary with key insights
-				var fetcherInfo string
-				if response.Processing.LogFetcher == "AgentLogFetcher" {
-					fetcherInfo = "⚡ AgentLogFetcher (Server-side processing)"
-					if response.Processing.ServerSideGrep {
-						fetcherInfo += " with server-side filtering"
-					}
-				} else {
-					fetcherInfo = "KubeLogFetcher (Kubernetes API - client-side processing)"
-				}
-
-				var summaryText strings.Builder
-
-				fmt.Fprintf(&summaryText, "Logs for query: \"%s\" (Found %d entries)\n\n",
-					query, response.Summary.TotalEntries)
-
-				fmt.Fprintf(&summaryText, "**Log Entries:**\n")
-
-				// Prioritize error and warning messages first
-				var errorLogs, warningLogs, otherLogs []LogEntryResponse
-				for _, entry := range response.Entries {
-					if entry.Level == "error" {
-						errorLogs = append(errorLogs, entry)
-					} else if entry.Level == "warning" {
-						warningLogs = append(warningLogs, entry)
-					} else {
-						otherLogs = append(otherLogs, entry)
-					}
-				}
-
-				// limit total entries to display to 10 entries
-				remainingEntries := 10
-
-				// errors first
-				if len(errorLogs) > 0 {
-					fmt.Fprintf(&summaryText, "\n **ERROR LOGS:**\n")
-					for _, entry := range errorLogs {
-						if remainingEntries <= 0 {
-							break
-						}
-						fmt.Fprintf(&summaryText, "  • [%s] %s/%s (%s):\n    %s\n\n",
-							entry.Timestamp,
-							entry.Source.Namespace,
-							entry.Source.PodName,
-							entry.Source.ContainerName,
-							entry.Message)
-						remainingEntries--
-					}
-				}
-
-				// warnings
-				if len(warningLogs) > 0 && remainingEntries > 0 {
-					fmt.Fprintf(&summaryText, "\n **WARNING LOGS:**\n")
-					for _, entry := range warningLogs {
-						if remainingEntries <= 0 {
-							break
-						}
-						fmt.Fprintf(&summaryText, "  • [%s] %s/%s (%s):\n    %s\n\n",
-							entry.Timestamp,
-							entry.Source.Namespace,
-							entry.Source.PodName,
-							entry.Source.ContainerName,
-							entry.Message)
-						remainingEntries--
-					}
-				}
-
-				// other logs
-				if len(otherLogs) > 0 && remainingEntries > 0 {
-					fmt.Fprintf(&summaryText, "\n **INFO LOGS:**\n")
-					for _, entry := range otherLogs {
-						if remainingEntries <= 0 {
-							break
-						}
-						fmt.Fprintf(&summaryText, "  • [%s] %s/%s (%s):\n    %s\n\n",
-							entry.Timestamp,
-							entry.Source.Namespace,
-							entry.Source.PodName,
-							entry.Source.ContainerName,
-							entry.Message)
-						remainingEntries--
-					}
-				}
-
-				// indicate there are more entries
-				if len(response.Entries) > 10 {
-					fmt.Fprintf(&summaryText, "\n Showing 10 of %d entries. Use format='raw' to see all logs.\n",
-						len(response.Entries))
-				}
-
-				// Display patterns and insights from log analysis
-				fmt.Fprintf(&summaryText, "\n **Log Analysis:**\n")
-
-				// insights about log patterns
-				if response.Summary.ErrorCount > 0 {
-					fmt.Fprintf(&summaryText, "• Found %d error messages - potential issues detected\n",
-						response.Summary.ErrorCount)
-				} else {
-					fmt.Fprintf(&summaryText, "• No errors detected in logs\n")
-				}
-
-				if response.Summary.WarningCount > 0 {
-					fmt.Fprintf(&summaryText, "• Found %d warning messages\n",
-						response.Summary.WarningCount)
-				}
-
-				fmt.Fprintf(&summaryText, "• Timespan: %s to %s\n",
-					response.Summary.TimeRange.Start,
-					response.Summary.TimeRange.End)
-
-				// show top keywords
-				if len(response.Summary.TopKeywords) > 0 {
-					fmt.Fprintf(&summaryText, "• Key topics: %s\n",
-						strings.Join(response.Summary.TopKeywords, ", "))
-				}
-
-				// source information
-				fmt.Fprintf(&summaryText, "\n**Sources:**\n")
-
-				// List all namespaces
-				if len(response.Summary.SourceSummary.Namespaces) > 0 {
-					fmt.Fprintf(&summaryText, "• Namespaces: %s\n",
-						strings.Join(response.Summary.SourceSummary.Namespaces, ", "))
-				}
-
-				// List all pods
-				if len(response.Summary.SourceSummary.Pods) > 0 {
-					if len(response.Summary.SourceSummary.Pods) <= 5 {
-						fmt.Fprintf(&summaryText, "• Pods: %s\n",
-							strings.Join(response.Summary.SourceSummary.Pods, ", "))
-					} else {
-						fmt.Fprintf(&summaryText, "• Pods: %d pods including %s\n",
-							len(response.Summary.SourceSummary.Pods),
-							strings.Join(response.Summary.SourceSummary.Pods[:3], ", "))
-					}
-				}
-
-				// other meta info
-				fmt.Fprintf(&summaryText, "\n **Technical Details:**\n")
-				fmt.Fprintf(&summaryText, "• Data source: %s\n", fetcherInfo)
-				fmt.Fprintf(&summaryText, "• Processing time: %dms\n", response.Processing.ProcessingTimeMs)
-				fmt.Fprintf(&summaryText, "• Sources scanned: %d\n", processingInfo.SourcesScanned)
-
-				if response.InfraInsights.InfrastructureSpan != "" {
-					fmt.Fprintf(&summaryText, "• Infrastructure: %s\n",
-						response.InfraInsights.InfrastructureSpan)
-				}
-
-				// applied filters
-				if appliedFilters.Namespace != "" || appliedFilters.Pod != "" ||
-					appliedFilters.Container != "" || appliedFilters.Grep != "" ||
-					len(appliedFilters.Region) > 0 || len(appliedFilters.Zone) > 0 ||
-					len(appliedFilters.Node) > 0 || appliedFilters.Since != "" ||
-					appliedFilters.Until != "" {
-
-					fmt.Fprintf(&summaryText, "\n**Applied Filters:**\n")
-
-					if appliedFilters.Namespace != "" {
-						fmt.Fprintf(&summaryText, "• Namespace: %s\n", appliedFilters.Namespace)
-					}
-					if appliedFilters.Pod != "" {
-						fmt.Fprintf(&summaryText, "• Pod: %s\n", appliedFilters.Pod)
-					}
-					if appliedFilters.Container != "" {
-						fmt.Fprintf(&summaryText, "• Container: %s\n", appliedFilters.Container)
-					}
-					if appliedFilters.Grep != "" {
-						fmt.Fprintf(&summaryText, "• Grep pattern: %s\n", appliedFilters.Grep)
-					}
-					if len(appliedFilters.Region) > 0 {
-						fmt.Fprintf(&summaryText, "• Regions: %s\n", strings.Join(appliedFilters.Region, ", "))
-					}
-					if len(appliedFilters.Zone) > 0 {
-						fmt.Fprintf(&summaryText, "• Zones: %s\n", strings.Join(appliedFilters.Zone, ", "))
-					}
-					if len(appliedFilters.Node) > 0 {
-						fmt.Fprintf(&summaryText, "• Nodes: %s\n", strings.Join(appliedFilters.Node, ", "))
-					}
-					if appliedFilters.Since != "" {
-						fmt.Fprintf(&summaryText, "• Since: %s\n", appliedFilters.Since)
-					}
-					if appliedFilters.Until != "" {
-						fmt.Fprintf(&summaryText, "• Until: %s\n", appliedFilters.Until)
-					}
-				}
-
-				// streaming instructions if follow is enabled
-				if follow {
-					fmt.Fprintf(&summaryText, "\n **Streaming Instructions:**\n")
-					fmt.Fprintf(&summaryText, "• To continue streaming logs, call again with since=\"%s\"\n",
-						response.Summary.TimeRange.End)
-				}
-
-				fmt.Fprintf(&summaryText, "\n **Note:** Use format='detailed' for full JSON data or format='raw' for traditional log lines.\n\n")
-
-				return mcp.NewToolResultText(summaryText.String()), nil
-			}
-		})
+		s.AddTool(tool, handleKubernetesLogSearchRequest(cm, kubeContext))
 
 		log.Info().Msg("MCP Server initialized, starting stdio transport...")
 
@@ -1008,453 +157,512 @@ var mcpServerCmd = &cobra.Command{
 	},
 }
 
-// buildSourcePaths creates log source paths based on namespace, pod, and container criteria.
-func buildSourcePaths(namespace, pod, container string, cm k8shelpers.ConnectionManager, kubeContext string) ([]string, error) {
-	var sourcePaths []string
+// creates a handler function for the kubernetes_logs_search MCP tool
+func handleKubernetesLogSearchRequest(cm k8shelpers.ConnectionManager, kubeContext string) func(ctx context.Context, request mcp.CallToolRequest) (*mcp.CallToolResult, error) {
+	return func(ctx context.Context, request mcp.CallToolRequest) (*mcp.CallToolResult, error) {
+		log.Info().Str("tool", "kubernetes_logs_search").Msg("Processing log search request")
 
-	// If no namespace specified, use default
-	if namespace == "" {
-		namespace = cm.GetDefaultNamespace(kubeContext)
+		args, ok := request.Params.Arguments.(map[string]interface{})
+		if !ok {
+			return mcp.NewToolResultError("invalid arguments format"), nil
+		}
+
+		query, ok := args["query"].(string)
+		if !ok || query == "" {
+			return mcp.NewToolResultError("query parameter is required"), nil
+		}
+
+		// Kubernetes filters
+		namespace, _ := args["namespace"].(string)
+		pod, _ := args["pod"].(string)
+		container, _ := args["container"].(string)
+
+		// Infrastructure filters
+		region, _ := args["region"].(string)
+		zone, _ := args["zone"].(string)
+		arch, _ := args["arch"].(string)
+		osFilter, _ := args["os"].(string)
+		node, _ := args["node"].(string)
+
+		// Content and time filters
+		grep, _ := args["grep"].(string)
+		since, _ := args["since"].(string)
+		until, _ := args["until"].(string)
+
+		// Output control
+		tail, _ := args["tail"].(float64)
+		follow, _ := args["follow"].(bool)
+		format, _ := args["format"].(string)
+
+		// defaults
+		if tail == 0 {
+			tail = 30 // default
+		}
+		if format == "" {
+			format = "detailed" // Default to structured JSON for AI consumption
+		}
+
+		sourcePaths, err := buildSourcePaths(namespace, pod, container, cm, kubeContext)
+		if err != nil {
+			return mcp.NewToolResultError(fmt.Sprintf("Failed to build source paths: %v", err)), nil
+		}
+
+		if len(sourcePaths) == 0 {
+			return mcp.NewToolResultText("No matching pods found for the specified criteria"), nil
+		}
+
+		streamOpts := []logs.Option{
+			logs.WithKubeContext(kubeContext),
+			logs.WithTail(int64(tail)),
+			logs.WithFollow(follow),
+		}
+
+		if region != "" {
+			regionList := parseCommaList(region)
+			streamOpts = append(streamOpts, logs.WithRegions(regionList))
+		}
+		if zone != "" {
+			zoneList := parseCommaList(zone)
+			streamOpts = append(streamOpts, logs.WithZones(zoneList))
+		}
+		if arch != "" {
+			archList := parseCommaList(arch)
+			streamOpts = append(streamOpts, logs.WithArches(archList))
+		}
+		if osFilter != "" {
+			osList := parseCommaList(osFilter)
+			streamOpts = append(streamOpts, logs.WithOSes(osList))
+		}
+		if node != "" {
+			nodeList := parseCommaList(node)
+			streamOpts = append(streamOpts, logs.WithNodes(nodeList))
+		}
+
+		if grep != "" {
+			streamOpts = append(streamOpts, logs.WithGrep(grep))
+		}
+
+		if since != "" {
+			sinceTime, err := parseTimeArgMCP(since)
+			if err != nil {
+				return mcp.NewToolResultError(fmt.Sprintf("Invalid since time: %v", err)), nil
+			}
+			streamOpts = append(streamOpts, logs.WithSince(sinceTime))
+		}
+		if until != "" {
+			untilTime, err := parseTimeArgMCP(until)
+			if err != nil {
+				return mcp.NewToolResultError(fmt.Sprintf("Invalid until time: %v", err)), nil
+			}
+			streamOpts = append(streamOpts, logs.WithUntil(untilTime))
+		}
+
+		// Initialize SmartLogFetcher
+		logFetcher, logFetcherType, err := createSmartLogFetcher(cm, kubeContext)
+		if err != nil {
+			return mcp.NewToolResultError(fmt.Sprintf("Failed to create log fetcher: %s", err)), nil
+		}
+		streamOpts = append(streamOpts, logs.WithLogFetcher(logFetcher))
+
+		startTime := time.Now()
+		stream, err := logs.NewStream(ctx, cm, sourcePaths, streamOpts...)
+		if err != nil {
+			return mcp.NewToolResultError(fmt.Sprintf("Failed to create log stream: %v", err)), nil
+		}
+		defer stream.Close()
+
+		if err := stream.Start(ctx); err != nil {
+			return mcp.NewToolResultError(fmt.Sprintf("Failed to start log stream: %v", err)), nil
+		}
+
+		processingTime := time.Since(startTime).Milliseconds()
+
+		appliedFilters := AppliedFilters{
+			Namespace: namespace,
+			Pod:       pod,
+			Container: container,
+			Grep:      grep,
+			Since:     since,
+			Until:     until,
+			Tail:      int64(tail),
+		}
+
+		if region != "" {
+			appliedFilters.Region = parseCommaList(region)
+		}
+		if zone != "" {
+			appliedFilters.Zone = parseCommaList(zone)
+		}
+		if arch != "" {
+			appliedFilters.Arch = parseCommaList(arch)
+		}
+		if osFilter != "" {
+			appliedFilters.OS = parseCommaList(osFilter)
+		}
+		if node != "" {
+			appliedFilters.Node = parseCommaList(node)
+		}
+
+		processingInfo := ProcessingInfo{
+			LogFetcher:       logFetcherType,
+			ServerSideGrep:   grep != "" && logFetcherType == "AgentLogFetcher",
+			ProcessingTimeMs: processingTime,
+			SourcesScanned:   len(sourcePaths),
+		}
+
+		// Handle streaming (follow) requests
+		if follow {
+			return handleStreamingRequest(ctx, stream, query, sourcePaths, streamOpts, since,
+				appliedFilters, processingInfo)
+		}
+
+		// For non-streaming requests, collect all records
+		var records []logs.LogRecord
+		for record := range stream.Records() {
+			records = append(records, record)
+		}
+
+		if stream.Err() != nil {
+			return mcp.NewToolResultError(fmt.Sprintf("Error while streaming logs: %v", stream.Err())), nil
+		}
+
+		response := buildStructuredResponse(query, records, appliedFilters, processingInfo)
+
+		if len(records) == 0 {
+			return mcp.NewToolResultText("No log entries found matching the specified criteria"), nil
+		}
+
+		return formatResponse(response, format, records)
 	}
+}
 
-	var sourcePath string
-	if pod != "" {
-		if container != "" {
-			sourcePath = fmt.Sprintf("%s:%s/%s", namespace, pod, container)
-		} else {
-			sourcePath = fmt.Sprintf("%s:%s", namespace, pod)
+// handles follow=true requests with a timeout-based approach
+func handleStreamingRequest(ctx context.Context, stream *logs.Stream, query string,
+	sourcePaths []string, streamOpts []logs.Option, since string,
+	appliedFilters AppliedFilters, processingInfo ProcessingInfo) (*mcp.CallToolResult, error) {
+
+	var streamingSince time.Time
+
+	// since parameter was provided
+	if since != "" {
+		var err error
+		streamingSince, err = parseTimeArgMCP(since)
+		if err != nil {
+			return mcp.NewToolResultError(fmt.Sprintf("Invalid since time for streaming: %v", err)), nil
 		}
 	} else {
-		if container != "" {
-			sourcePath = fmt.Sprintf("%s:*/%s", namespace, container)
-		} else {
-			sourcePath = fmt.Sprintf("%s:*", namespace)
+		// no since parameter, use current time, start collecting from 5 seconds ago
+		streamingSince = time.Now().Add(-5 * time.Second)
+	}
+
+	hasSinceOption := false
+	for i, opt := range streamOpts {
+		// check the option, if it's a WithSince
+		if optStr := fmt.Sprintf("%T", opt); strings.Contains(optStr, "WithSince") {
+			streamOpts[i] = logs.WithSince(streamingSince)
+			hasSinceOption = true
+			break
 		}
 	}
 
-	sourcePaths = append(sourcePaths, sourcePath)
-	return sourcePaths, nil
-}
-
-/* Utility functions */
-
-// parseCommaList splits a comma-separated string into a slice, trimming whitespace.
-func parseCommaList(input string) []string {
-	if input == "" {
-		return nil
+	// add since option if not found
+	if !hasSinceOption {
+		streamOpts = append(streamOpts, logs.WithSince(streamingSince))
 	}
 
-	var result []string
-	for _, item := range strings.Split(input, ",") {
-		trimmed := strings.TrimSpace(item)
-		if trimmed != "" {
-			result = append(result, trimmed)
-		}
-	}
-	return result
-}
+	var initialRecords []logs.LogRecord
+	recordCount := 0
+	recordLimit := 20 // initial batch size
 
-// parseTimeArgMCP parses time arguments for the MCP server, supporting ISO durations and timestamps.
-func parseTimeArgMCP(arg string) (time.Time, error) {
-	var zero time.Time
+	timeout := time.After(2 * time.Second)
+	collectDone := false
 
-	arg = strings.TrimSpace(arg)
-	if arg == "" {
-		return zero, nil
-	} else if timeAgo, err := duration.Parse(arg); err == nil {
-		// Parsed as ISO duration
-		return time.Now().Add(-1 * timeAgo.ToTimeDuration()), nil
-	} else if ts, err := time.Parse(time.RFC3339Nano, arg); err == nil {
-		// Parsed as ISO timestamp
-		return ts, nil
-	}
-
-	return zero, fmt.Errorf("unable to parse arg %s", arg)
-}
-
-// detectLogLevel attempts to infer log level from message content.
-func detectLogLevel(message string) string {
-	lower := strings.ToLower(message)
-
-	if strings.Contains(lower, "error") || strings.Contains(lower, "err") || strings.Contains(lower, "fatal") {
-		return "error"
-	}
-	if strings.Contains(lower, "warn") || strings.Contains(lower, "warning") {
-		return "warning"
-	}
-	if strings.Contains(lower, "info") || strings.Contains(lower, "information") {
-		return "info"
-	}
-	if strings.Contains(lower, "debug") || strings.Contains(lower, "trace") {
-		return "debug"
-	}
-
-	// Check for HTTP status codes, 4xx/5xx status
-	if strings.Contains(message, " 4") || strings.Contains(message, " 5") {
-		return "error"
-	}
-	if strings.Contains(message, " 2") || strings.Contains(message, " 3") { // 2xx/3xx status
-		return "info"
-	}
-
-	return "info" // Default level
-}
-
-// extractTags identifies relevant tags/labels from log messages for categorization.
-func extractTags(message string) []string {
-	var tags []string
-	lower := strings.ToLower(message)
-
-	// HTTP-related tags
-	if strings.Contains(lower, "http") || strings.Contains(message, "GET") || strings.Contains(message, "POST") {
-		tags = append(tags, "http")
-	}
-
-	// Database-related tags
-	if strings.Contains(lower, "sql") || strings.Contains(lower, "database") || strings.Contains(lower, "db") {
-		tags = append(tags, "database")
-	}
-
-	// Authentication/authorization
-	if strings.Contains(lower, "auth") || strings.Contains(lower, "login") || strings.Contains(lower, "permission") {
-		tags = append(tags, "authentication")
-	}
-
-	// API-related
-	if strings.Contains(lower, "api") || strings.Contains(lower, "/v1/") || strings.Contains(lower, "/v2/") {
-		tags = append(tags, "api")
-	}
-
-	// Performance-related
-	if strings.Contains(lower, "timeout") || strings.Contains(lower, "slow") || strings.Contains(lower, "latency") {
-		tags = append(tags, "performance")
-	}
-
-	// Network-related
-	if strings.Contains(lower, "connection") || strings.Contains(lower, "network") || strings.Contains(lower, "tcp") {
-		tags = append(tags, "network")
-	}
-
-	return tags
-}
-
-// extractLogMetadata extracts structured metadata from log messages for analysis.
-func extractLogMetadata(message, level string) map[string]string {
-	metadata := make(map[string]string)
-
-	if strings.Contains(message, " 200 ") || strings.Contains(message, " 201 ") {
-		metadata["http_status"] = "success"
-		metadata["http_status_code"] = "2xx"
-	} else if strings.Contains(message, " 4") && (strings.Contains(message, " 40") || strings.Contains(message, " 41")) {
-		metadata["http_status"] = "client_error"
-		metadata["http_status_code"] = "4xx"
-	} else if strings.Contains(message, " 5") && strings.Contains(message, " 50") {
-		metadata["http_status"] = "server_error"
-		metadata["http_status_code"] = "5xx"
-	}
-
-	if strings.Contains(message, "GET ") {
-		metadata["http_method"] = "GET"
-	} else if strings.Contains(message, "POST ") {
-		metadata["http_method"] = "POST"
-	} else if strings.Contains(message, "PUT ") {
-		metadata["http_method"] = "PUT"
-	} else if strings.Contains(message, "DELETE ") {
-		metadata["http_method"] = "DELETE"
-	}
-
-	if strings.Contains(message, ".") {
-		words := strings.Fields(message)
-		for _, word := range words {
-			if strings.Count(word, ".") == 3 && len(word) >= 7 {
-				metadata["client_ip"] = word
+	for !collectDone {
+		select {
+		case record, ok := <-stream.Records():
+			if !ok {
+				collectDone = true
 				break
 			}
-		}
-	}
-
-	if strings.Contains(message, "ms") || strings.Contains(message, "seconds") {
-		metadata["contains_timing"] = "true"
-	}
-
-	if strings.Contains(message, "{") && strings.Contains(message, "}") {
-		metadata["log_format"] = "json"
-	} else if strings.Count(message, "|") > 2 || strings.Count(message, ",") > 3 {
-		metadata["log_format"] = "structured"
-	} else {
-		metadata["log_format"] = "unstructured"
-	}
-
-	metadata["severity"] = level
-
-	return metadata
-}
-
-// extractKeywords analyzes a slice of log messages and returns the most frequent, meaningful keywords.
-// This is used to help summarize and surface the main topics or issues present in the logs,
-// making it easier for both AI tools and human users to quickly understand the context of log search results.
-func extractKeywords(messages []string) []string {
-	wordCount := make(map[string]int)
-	commonWords := map[string]bool{
-		"the": true, "and": true, "or": true, "but": true, "in": true, "on": true, "at": true,
-		"to": true, "for": true, "of": true, "with": true, "by": true, "is": true, "are": true,
-		"was": true, "were": true, "be": true, "been": true, "have": true, "has": true, "had": true,
-		"do": true, "does": true, "did": true, "will": true, "would": true, "could": true, "should": true,
-		"a": true, "an": true, "this": true, "that": true, "these": true, "those": true,
-	}
-
-	for _, message := range messages {
-		words := strings.Fields(strings.ToLower(message))
-		for _, word := range words {
-			// remove punctuation
-			cleaned := strings.Trim(word, ".,!?:;\"'()[]{}/-")
-
-			// Skip common words and short words
-			if len(cleaned) < 3 || commonWords[cleaned] {
-				continue
+			initialRecords = append(initialRecords, record)
+			recordCount++
+			if recordCount >= recordLimit {
+				collectDone = true
 			}
-
-			wordCount[cleaned]++
+		case <-timeout:
+			collectDone = true
+		case <-ctx.Done():
+			return mcp.NewToolResultError("Request cancelled"), nil
 		}
 	}
 
-	type wordFreq struct {
-		word  string
-		count int
-	}
-
-	var frequencies []wordFreq
-	for word, count := range wordCount {
-		if count > 1 {
-			frequencies = append(frequencies, wordFreq{word, count})
-		}
-	}
-
-	// Sort by frequency
-	for i := 0; i < len(frequencies)-1; i++ {
-		for j := i + 1; j < len(frequencies); j++ {
-			if frequencies[j].count > frequencies[i].count {
-				frequencies[i], frequencies[j] = frequencies[j], frequencies[i]
+	// get latest timestamp from logs for next polling request
+	var latestTimestamp time.Time
+	if len(initialRecords) > 0 {
+		for _, record := range initialRecords {
+			if record.Timestamp.After(latestTimestamp) {
+				latestTimestamp = record.Timestamp
 			}
 		}
+	} else {
+		latestTimestamp = time.Now()
 	}
 
-	// top 5 keywords
-	var keywords []string
-	limit := 5
-	if len(frequencies) < limit {
-		limit = len(frequencies)
+	// 1 millisecond increment to prevent duplicate logs in next poll
+	nextPollTime := latestTimestamp.Add(time.Millisecond)
+
+	response := buildStructuredResponse(query, initialRecords, appliedFilters, processingInfo)
+
+	// meta info
+	if response.InfraInsights.AppliedFilters == nil {
+		response.InfraInsights.AppliedFilters = []string{}
+	}
+	response.InfraInsights.AppliedFilters = append(
+		response.InfraInsights.AppliedFilters,
+		"Real-time log streaming requires repeated calls with the same parameters")
+
+	if len(initialRecords) == 0 {
+		response.Summary.TotalEntries = 0
+		jsonData, err := json.MarshalIndent(response, "", "  ")
+		if err != nil {
+			return mcp.NewToolResultError(fmt.Sprintf("Failed to serialize response: %v", err)), nil
+		}
+
+		// Use current time as the next poll point
+		nextPollStr := time.Now().Format(time.RFC3339)
+		return mcp.NewToolResultText(fmt.Sprintf("No log entries found. For continuous updates, call again with: since=\"%s\"\n\n%s",
+			nextPollStr, string(jsonData))), nil
 	}
 
-	for i := 0; i < limit; i++ {
-		keywords = append(keywords, frequencies[i].word)
+	jsonData, err := json.MarshalIndent(response, "", "  ")
+	if err != nil {
+		return mcp.NewToolResultError(fmt.Sprintf("Failed to serialize response: %v", err)), nil
 	}
 
-	return keywords
+	nextPollStr := nextPollTime.Format(time.RFC3339)
+
+	return mcp.NewToolResultText(fmt.Sprintf("Found %d log entries. For continuous updates, call again with: since=\"%s\"\n\n%s",
+		len(initialRecords), nextPollStr, string(jsonData))), nil
 }
 
-// buildStructuredResponse creates a structured response for log search results, including summary and insights.
-func buildStructuredResponse(query string, records []logs.LogRecord, filters AppliedFilters, processingInfo ProcessingInfo) *LogSearchResponse {
-	if len(records) == 0 {
-		InfraInsights := buildInfrastructureInsights(filters, processingInfo)
-		return &LogSearchResponse{
-			Query: query,
-			Summary: LogSummary{
-				TotalEntries: 0,
-			},
-			Entries:       []LogEntryResponse{},
-			Filters:       filters,
-			Processing:    processingInfo,
-			InfraInsights: InfraInsights,
+
+func formatResponse(response *LogSearchResponse, format string, records []logs.LogRecord) (*mcp.CallToolResult, error) {
+	switch format {
+	case "raw":
+		var logLines []string
+		for _, record := range records {
+			logLine := fmt.Sprintf("[%s] %s/%s/%s: %s",
+				record.Timestamp.Format("2006-01-02 15:04:05"),
+				record.Source.Namespace,
+				record.Source.PodName,
+				record.Source.ContainerName,
+				record.Message)
+			logLines = append(logLines, logLine)
 		}
-	}
+		result := fmt.Sprintf("Found %d log entries:\n\n%s", len(logLines), strings.Join(logLines, "\n"))
+		return mcp.NewToolResultText(result), nil
 
-	var entries []LogEntryResponse
-	var messages []string
-	var errorCount, warningCount int
-	levelCounts := make(map[string]int)
-	namespaces := make(map[string]bool)
-	pods := make(map[string]bool)
-	containers := make(map[string]bool)
-
-	var earliestTime, latestTime time.Time
-
-	for i, record := range records {
-		level := detectLogLevel(record.Message)
-
-		metadata := extractLogMetadata(record.Message, level)
-
-		entry := LogEntryResponse{
-			Timestamp: record.Timestamp.Format(time.RFC3339),
-			Message:   record.Message,
-			Source:    record.Source,
-			Level:     level,
-			Tags:      extractTags(record.Message),
-			Metadata:  metadata,
+	case "detailed":
+		jsonData, err := json.MarshalIndent(response, "", "  ")
+		if err != nil {
+			return mcp.NewToolResultError(fmt.Sprintf("Failed to serialize structured response: %v", err)), nil
 		}
+		return mcp.NewToolResultText(fmt.Sprintf("Structured log data (JSON):\n\n%s", string(jsonData))), nil
 
-		entries = append(entries, entry)
-		messages = append(messages, record.Message)
-
-		levelCounts[level]++
-		if level == "error" {
-			errorCount++
-		}
-		if level == "warning" {
-			warningCount++
-		}
-
-		namespaces[record.Source.Namespace] = true
-		pods[record.Source.PodName] = true
-		containers[record.Source.ContainerName] = true
-
-		if i == 0 || record.Timestamp.Before(earliestTime) {
-			earliestTime = record.Timestamp
-		}
-		if i == 0 || record.Timestamp.After(latestTime) {
-			latestTime = record.Timestamp
-		}
-	}
-
-	summary := LogSummary{
-		TotalEntries: len(records),
-		TimeRange: TimeRange{
-			Start: earliestTime.Format(time.RFC3339),
-			End:   latestTime.Format(time.RFC3339),
-		},
-		SourceSummary: SourceSummary{
-			Namespaces: mapKeysToSlice(namespaces),
-			Pods:       mapKeysToSlice(pods),
-			Containers: mapKeysToSlice(containers),
-		},
-		LevelCounts:  levelCounts,
-		TopKeywords:  extractKeywords(messages),
-		ErrorCount:   errorCount,
-		WarningCount: warningCount,
-	}
-
-	InfraInsights := buildInfrastructureInsights(filters, processingInfo)
-
-	return &LogSearchResponse{
-		Query:         query,
-		Summary:       summary,
-		Entries:       entries,
-		Filters:       filters,
-		Processing:    processingInfo,
-		InfraInsights: InfraInsights,
+	default: // "summary"
+		return formatSummaryResponse(response), nil
 	}
 }
 
-// mapKeysToSlice converts a map[string]bool to []string
-func mapKeysToSlice(m map[string]bool) []string {
-	var result []string
-	for key := range m {
-		result = append(result, key)
-	}
-	return result
-}
-
-// buildInfrastructureInsights summarizes Kubetail's infrastructure-aware capabilities for the response.
-func buildInfrastructureInsights(filters AppliedFilters, processing ProcessingInfo) InfrastructureInsights {
-	var appliedFiltersList []string
-	var infrastructureSpan string
-	var processingMethod string
-
-	// Detect advanced infrastructure filtering
-	advancedFiltering := len(filters.Region) > 0 || len(filters.Zone) > 0 ||
-		len(filters.Arch) > 0 || len(filters.OS) > 0 || len(filters.Node) > 0
-
-	// Determine server-side processing
-	serverSideFiltering := processing.LogFetcher == "AgentLogFetcher"
-
-	// Collect applied filters
-	if len(filters.Region) > 0 {
-		appliedFiltersList = append(appliedFiltersList, "Multi-region log aggregation")
-		infrastructureSpan += fmt.Sprintf("Regions: %v ", filters.Region)
-	}
-	if len(filters.Zone) > 0 {
-		appliedFiltersList = append(appliedFiltersList, "Availability zone filtering")
-		infrastructureSpan += fmt.Sprintf("Zones: %v ", filters.Zone)
-	}
-	if len(filters.Arch) > 0 {
-		appliedFiltersList = append(appliedFiltersList, "CPU architecture-aware filtering")
-		infrastructureSpan += fmt.Sprintf("Architectures: %v ", filters.Arch)
-	}
-	if len(filters.OS) > 0 {
-		appliedFiltersList = append(appliedFiltersList, "Operating system filtering")
-		infrastructureSpan += fmt.Sprintf("OS: %v ", filters.OS)
-	}
-	if len(filters.Node) > 0 {
-		appliedFiltersList = append(appliedFiltersList, "Node-specific targeting")
-		infrastructureSpan += fmt.Sprintf("Nodes: %v ", filters.Node)
-	}
-
-	if serverSideFiltering {
-		appliedFiltersList = append(appliedFiltersList, "Agent-based server-side log processing")
-		processingMethod = "Logs filtered at source by cluster agents, reducing network transfer and improving query performance"
-		if processing.ServerSideGrep {
-			appliedFiltersList = append(appliedFiltersList, "Server-side regex filtering")
+func formatSummaryResponse(response *LogSearchResponse) *mcp.CallToolResult {
+	var fetcherInfo string
+	if response.Processing.LogFetcher == "AgentLogFetcher" {
+		fetcherInfo = "⚡ AgentLogFetcher (Server-side processing)"
+		if response.Processing.ServerSideGrep {
+			fetcherInfo += " with server-side filtering"
 		}
 	} else {
-		processingMethod = "Direct Kubernetes API access with client-side processing"
+		fetcherInfo = "KubeLogFetcher (Kubernetes API - client-side processing)"
 	}
 
-	if infrastructureSpan == "" {
-		infrastructureSpan = "Single namespace/cluster scope"
+	var summaryText strings.Builder
+
+	fmt.Fprintf(&summaryText, "Logs for query: \"%s\" (Found %d entries)\n\n",
+		response.Query, response.Summary.TotalEntries)
+
+	fmt.Fprintf(&summaryText, "**Log Entries:**\n")
+
+	// prioritize error and warning messages first
+	var errorLogs, warningLogs, otherLogs []LogEntryResponse
+	for _, entry := range response.Entries {
+		if entry.Level == "error" {
+			errorLogs = append(errorLogs, entry)
+		} else if entry.Level == "warning" {
+			warningLogs = append(warningLogs, entry)
+		} else {
+			otherLogs = append(otherLogs, entry)
+		}
+	}
+
+	// limit total entries to display 10 entries
+	remainingEntries := 10
+
+	// errors first
+	if len(errorLogs) > 0 {
+		fmt.Fprintf(&summaryText, "\n **ERROR LOGS:**\n")
+		for _, entry := range errorLogs {
+			if remainingEntries <= 0 {
+				break
+			}
+			fmt.Fprintf(&summaryText, "  • [%s] %s/%s (%s):\n    %s\n\n",
+				entry.Timestamp,
+				entry.Source.Namespace,
+				entry.Source.PodName,
+				entry.Source.ContainerName,
+				entry.Message)
+			remainingEntries--
+		}
+	}
+
+	// warnings
+	if len(warningLogs) > 0 && remainingEntries > 0 {
+		fmt.Fprintf(&summaryText, "\n **WARNING LOGS:**\n")
+		for _, entry := range warningLogs {
+			if remainingEntries <= 0 {
+				break
+			}
+			fmt.Fprintf(&summaryText, "  • [%s] %s/%s (%s):\n    %s\n\n",
+				entry.Timestamp,
+				entry.Source.Namespace,
+				entry.Source.PodName,
+				entry.Source.ContainerName,
+				entry.Message)
+			remainingEntries--
+		}
+	}
+
+	// other logs
+	if len(otherLogs) > 0 && remainingEntries > 0 {
+		fmt.Fprintf(&summaryText, "\n **INFO LOGS:**\n")
+		for _, entry := range otherLogs {
+			if remainingEntries <= 0 {
+				break
+			}
+			fmt.Fprintf(&summaryText, "  • [%s] %s/%s (%s):\n    %s\n\n",
+				entry.Timestamp,
+				entry.Source.Namespace,
+				entry.Source.PodName,
+				entry.Source.ContainerName,
+				entry.Message)
+			remainingEntries--
+		}
+	}
+
+	// indicate there are more entries
+	if len(response.Entries) > 10 {
+		fmt.Fprintf(&summaryText, "\n Showing 10 of %d entries. Use format='raw' to see all logs.\n",
+			len(response.Entries))
+	}
+
+	// Display patterns and insights from log analysis
+	fmt.Fprintf(&summaryText, "\n **Log Analysis:**\n")
+
+	// insights about log patterns
+	if response.Summary.ErrorCount > 0 {
+		fmt.Fprintf(&summaryText, "• Found %d error messages - potential issues detected\n",
+			response.Summary.ErrorCount)
 	} else {
-		infrastructureSpan = strings.TrimSpace(infrastructureSpan)
+		fmt.Fprintf(&summaryText, "• No errors detected in logs\n")
 	}
 
-	if filters.Grep != "" {
-		appliedFiltersList = append(appliedFiltersList, "Advanced regex pattern matching")
+	if response.Summary.WarningCount > 0 {
+		fmt.Fprintf(&summaryText, "• Found %d warning messages\n",
+			response.Summary.WarningCount)
 	}
 
-	appliedFiltersList = append(appliedFiltersList, "Real-time log streaming", "Multi-container correlation", "Structured metadata extraction")
+	fmt.Fprintf(&summaryText, "• Timespan: %s to %s\n",
+		response.Summary.TimeRange.Start,
+		response.Summary.TimeRange.End)
 
-	return InfrastructureInsights{
-		AdvancedFiltering:   advancedFiltering,
-		ServerSideFiltering: serverSideFiltering,
-		ProcessingMethod:    processingMethod,
-		AppliedFilters:      appliedFiltersList,
-		InfrastructureSpan:  infrastructureSpan,
-	}
-}
-
-// createSmartLogFetcher creates a LogFetcher with ClusterAPIProxy priority and graceful fallback
-// This is Kubetail's key advantage: server-side processing when cluster agents are available
-// Returns the fetcher and a string indicating which type was used ("AgentLogFetcher" or "KubeLogFetcher")
-func createSmartLogFetcher(cm k8shelpers.ConnectionManager, kubeContext string) (logs.LogFetcher, string, error) {
-
-	clientset, err := cm.GetOrCreateClientset(kubeContext)
-	if err != nil {
-		log.Error().Err(err).Msg("Failed to create Kubernetes clientset")
-		return nil, "KubeLogFetcher", err
+	// show top keywords
+	if len(response.Summary.TopKeywords) > 0 {
+		fmt.Fprintf(&summaryText, "• Key topics: %s\n",
+			strings.Join(response.Summary.TopKeywords, ", "))
 	}
 
-	// Try to create AgentLogFetcher via proxy
-	fetcher, fetcherType, err := tryCreateAgentLogFetcher(cm, kubeContext, clientset)
-	if err != nil {
-		log.Debug().Err(err).Msg("Failed to create AgentLogFetcher, using KubeLogFetcher")
-		return logs.NewKubeLogFetcher(clientset), "KubeLogFetcher", nil
+	// source information
+	fmt.Fprintf(&summaryText, "\n**Sources:**\n")
+
+	// List all namespaces
+	if len(response.Summary.SourceSummary.Namespaces) > 0 {
+		fmt.Fprintf(&summaryText, "• Namespaces: %s\n",
+			strings.Join(response.Summary.SourceSummary.Namespaces, ", "))
 	}
 
-	log.Info().Msgf("Using %s for server-side log processing", fetcherType)
-	return fetcher, fetcherType, nil
-}
-
-// tryCreateAgentLogFetcher attempts to create an ClusterAPIProxyLogFetcher (AgentLogFetcher), falling back KubeLogFetcher if unavailable.
-func tryCreateAgentLogFetcher(cm k8shelpers.ConnectionManager, kubeContext string, clientset kubernetes.Interface) (logs.LogFetcher, string, error) {
-	log.Debug().Msg("Using ClusterAPIProxyLogFetcher for accessing Kubetail agents")
-
-	fetcher, err := NewClusterAPIProxyLogFetcher(cm, kubeContext)
-	if err == nil {
-		log.Debug().Msg("Successfully created ClusterAPIProxyLogFetcher for AgentLogFetcher access")
-		return fetcher, "AgentLogFetcher", nil
+	// List all pods
+	if len(response.Summary.SourceSummary.Pods) > 0 {
+		if len(response.Summary.SourceSummary.Pods) <= 5 {
+			fmt.Fprintf(&summaryText, "• Pods: %s\n",
+				strings.Join(response.Summary.SourceSummary.Pods, ", "))
+		} else {
+			fmt.Fprintf(&summaryText, "• Pods: %d pods including %s\n",
+				len(response.Summary.SourceSummary.Pods),
+				strings.Join(response.Summary.SourceSummary.Pods[:3], ", "))
+		}
 	}
-	log.Debug().Err(err).Msg("Could not create ClusterAPIProxyLogFetcher, falling back to KubeLogFetcher (Kubernetes API)")
 
-	// Fallback to KubeLogFetcher
-	return logs.NewKubeLogFetcher(clientset), "KubeLogFetcher", nil
+	// other meta info
+	fmt.Fprintf(&summaryText, "\n **Technical Details:**\n")
+	fmt.Fprintf(&summaryText, "• Data source: %s\n", fetcherInfo)
+	fmt.Fprintf(&summaryText, "• Processing time: %dms\n", response.Processing.ProcessingTimeMs)
+	fmt.Fprintf(&summaryText, "• Sources scanned: %d\n", response.Processing.SourcesScanned)
+
+	if response.InfraInsights.InfrastructureSpan != "" {
+		fmt.Fprintf(&summaryText, "• Infrastructure: %s\n",
+			response.InfraInsights.InfrastructureSpan)
+	}
+
+	// applied filters
+	if response.Filters.Namespace != "" || response.Filters.Pod != "" ||
+		response.Filters.Container != "" || response.Filters.Grep != "" ||
+		len(response.Filters.Region) > 0 || len(response.Filters.Zone) > 0 ||
+		len(response.Filters.Node) > 0 || response.Filters.Since != "" ||
+		response.Filters.Until != "" {
+
+		fmt.Fprintf(&summaryText, "\n**Applied Filters:**\n")
+
+		if response.Filters.Namespace != "" {
+			fmt.Fprintf(&summaryText, "• Namespace: %s\n", response.Filters.Namespace)
+		}
+		if response.Filters.Pod != "" {
+			fmt.Fprintf(&summaryText, "• Pod: %s\n", response.Filters.Pod)
+		}
+		if response.Filters.Container != "" {
+			fmt.Fprintf(&summaryText, "• Container: %s\n", response.Filters.Container)
+		}
+		if response.Filters.Grep != "" {
+			fmt.Fprintf(&summaryText, "• Grep pattern: %s\n", response.Filters.Grep)
+		}
+		if len(response.Filters.Region) > 0 {
+			fmt.Fprintf(&summaryText, "• Regions: %s\n", strings.Join(response.Filters.Region, ", "))
+		}
+		if len(response.Filters.Zone) > 0 {
+			fmt.Fprintf(&summaryText, "• Zones: %s\n", strings.Join(response.Filters.Zone, ", "))
+		}
+		if len(response.Filters.Node) > 0 {
+			fmt.Fprintf(&summaryText, "• Nodes: %s\n", strings.Join(response.Filters.Node, ", "))
+		}
+		if response.Filters.Since != "" {
+			fmt.Fprintf(&summaryText, "• Since: %s\n", response.Filters.Since)
+		}
+		if response.Filters.Until != "" {
+			fmt.Fprintf(&summaryText, "• Until: %s\n", response.Filters.Until)
+		}
+	}
+
+	fmt.Fprintf(&summaryText, "\n **Note:** Use format='detailed' for full JSON data or format='raw' for traditional log lines.\n\n")
+
+	return mcp.NewToolResultText(summaryText.String())
 }
 
 func init() {
